Cargo related bug squashes in Cargo & Journal Monitors

+++ JournalMonitor/JournalMonitor.cs
@@ -1792,7 +1792,7 @@ namespace EddiJournalMonitor
                                 data.TryGetValue("Count", out val);
                                 int amount = (int)(long)val;
                                 data.TryGetValue("BuyPrice", out val);
-                                int price = (int)val;
+                                int price = (int)(long)val;
                                 events.Add(new LimpetPurchasedEvent(timestamp, amount, price) { raw = line });
                                 handled = true;
                                 break;
@@ -1803,7 +1803,7 @@ namespace EddiJournalMonitor
                                 data.TryGetValue("Count", out val);
                                 int amount = (int)(long)val;
                                 data.TryGetValue("SellPrice", out val);
-                                int price = (int)val;
+                                int price = (int)(long)val;
                                 events.Add(new LimpetSoldEvent(timestamp, amount, price) { raw = line });
                                 handled = true;
                                 break;
